Project/OCR Backend/requirements.txt
pytesseract==0.3.2
numpy==1.18.1
opencv_python==4.2.0.32
Flask==0.12.2
Pillow==7.1.2
